+++ SurgSim/Graphics/UnitTests/MeshTests.cpp
@@ -45,7 +45,7 @@ public:
 
 
 	std::vector<Vector3d> cubeVertices;
-	std::vector<unsigned int> cubeTriangles;
+	std::vector<size_t> cubeTriangles;
 	std::vector<Vector4d> cubeColors;
 	std::vector<Vector2d> cubeTextures;
 };
