+++ SurgSim/Physics/Fem3DElementTetrahedron.h
@@ -44,7 +44,7 @@ public:
 	/// \note This is required from the signed volume calculation method getVolume()
 	/// \note A warning will be logged when the initialize function is called if this condition is not met, but the
 	/// \note simulation will keep running.  Behavior will be undefined because of possible negative volume terms.
-	explicit Fem3DElementTetrahedron(std::array<unsigned int, 4> nodeIds);
+	explicit Fem3DElementTetrahedron(std::array<size_t, 4> nodeIds);
 
 	/// Initialize the FemElement once everything has been set
 	/// \param state The state to initialize the FemElement with
